+++ sdk/src/main/java/com/bugsnag/android/Client.java
@@ -143,8 +143,10 @@ public class Client extends Observable implements Observer {
         // Set up breadcrumbs
         breadcrumbs = new Breadcrumbs(configuration);
 
-        // Set sensible defaults
-        setProjectPackages(appContext.getPackageName());
+        // Set sensible defaults if project packages not already set
+        if (config.getProjectPackages() == null) {
+            setProjectPackages(appContext.getPackageName());
+        }
 
         String deviceId = deviceData.getId();
 
@@ -709,7 +711,7 @@ public class Client extends Observable implements Observer {
      * @param exception the exception to send to Bugsnag
      */
     public void notify(@NonNull Throwable exception) {
-        Error error = new Error.Builder(config, exception, sessionTracker.getCurrentSession(),
+        Error error = new Error.Builder(config, exception, sessionTracker,
             Thread.currentThread(), false)
             .severityReasonType(HandledState.REASON_HANDLED_EXCEPTION)
             .build();
@@ -724,7 +726,7 @@ public class Client extends Observable implements Observer {
      *                  additional modification
      */
     public void notify(@NonNull Throwable exception, @Nullable Callback callback) {
-        Error error = new Error.Builder(config, exception, sessionTracker.getCurrentSession(),
+        Error error = new Error.Builder(config, exception, sessionTracker,
             Thread.currentThread(), false)
             .severityReasonType(HandledState.REASON_HANDLED_EXCEPTION)
             .build();
@@ -745,7 +747,7 @@ public class Client extends Observable implements Observer {
                        @NonNull StackTraceElement[] stacktrace,
                        @Nullable Callback callback) {
         Error error = new Error.Builder(config, name, message, stacktrace,
-            sessionTracker.getCurrentSession(), Thread.currentThread())
+            sessionTracker, Thread.currentThread())
             .severityReasonType(HandledState.REASON_HANDLED_EXCEPTION)
             .build();
         notify(error, DeliveryStyle.ASYNC, callback);
@@ -759,7 +761,7 @@ public class Client extends Observable implements Observer {
      *                  Severity.WARNING or Severity.INFO
      */
     public void notify(@NonNull Throwable exception, @NonNull Severity severity) {
-        Error error = new Error.Builder(config, exception, sessionTracker.getCurrentSession(),
+        Error error = new Error.Builder(config, exception, sessionTracker,
             Thread.currentThread(), false)
             .severity(severity)
             .build();
@@ -776,7 +778,7 @@ public class Client extends Observable implements Observer {
     @Deprecated
     public void notify(@NonNull Throwable exception,
                        @NonNull MetaData metaData) {
-        Error error = new Error.Builder(config, exception, sessionTracker.getCurrentSession(),
+        Error error = new Error.Builder(config, exception, sessionTracker,
             Thread.currentThread(), false)
             .metaData(metaData)
             .severityReasonType(HandledState.REASON_HANDLED_EXCEPTION)
@@ -796,7 +798,7 @@ public class Client extends Observable implements Observer {
     @Deprecated
     public void notify(@NonNull Throwable exception, @NonNull Severity severity,
                        @NonNull MetaData metaData) {
-        Error error = new Error.Builder(config, exception, sessionTracker.getCurrentSession(),
+        Error error = new Error.Builder(config, exception, sessionTracker,
             Thread.currentThread(), false)
             .metaData(metaData)
             .severity(severity)
@@ -821,7 +823,7 @@ public class Client extends Observable implements Observer {
                        @NonNull StackTraceElement[] stacktrace, @NonNull Severity severity,
                        @NonNull MetaData metaData) {
         Error error = new Error.Builder(config, name, message,
-            stacktrace, sessionTracker.getCurrentSession(), Thread.currentThread())
+            stacktrace, sessionTracker, Thread.currentThread())
             .severity(severity)
             .metaData(metaData)
             .build();
@@ -849,7 +851,7 @@ public class Client extends Observable implements Observer {
                        @NonNull Severity severity,
                        @NonNull MetaData metaData) {
         Error error = new Error.Builder(config, name, message,
-            stacktrace, sessionTracker.getCurrentSession(), Thread.currentThread())
+            stacktrace, sessionTracker, Thread.currentThread())
             .severity(severity)
             .metaData(metaData)
             .build();
@@ -915,17 +917,10 @@ public class Client extends Observable implements Observer {
             callback.beforeNotify(report);
         }
 
-        HandledState handledState = report.getError().getHandledState();
-
-        if (handledState.isUnhandled()) {
-            sessionTracker.incrementUnhandledError();
-        } else {
-            sessionTracker.incrementHandledError();
-            if (sessionTracker.getCurrentSession() != null) {
-                setChanged();
-                notifyObservers(new NativeInterface.Message(
-                            NativeInterface.MessageType.NOTIFY_HANDLED, error.getExceptionName()));
-            }
+        if (!error.getHandledState().isUnhandled() && error.getSession() != null) {
+            setChanged();
+            notifyObservers(new NativeInterface.Message(
+                NativeInterface.MessageType.NOTIFY_HANDLED, error.getExceptionName()));
         }
 
         switch (style) {
@@ -971,7 +966,7 @@ public class Client extends Observable implements Observer {
      * @param exception the exception to send to Bugsnag
      */
     public void notifyBlocking(@NonNull Throwable exception) {
-        Error error = new Error.Builder(config, exception, sessionTracker.getCurrentSession(),
+        Error error = new Error.Builder(config, exception, sessionTracker,
             Thread.currentThread(), false)
             .severityReasonType(HandledState.REASON_HANDLED_EXCEPTION)
             .build();
@@ -986,7 +981,7 @@ public class Client extends Observable implements Observer {
      *                  additional modification
      */
     public void notifyBlocking(@NonNull Throwable exception, @Nullable Callback callback) {
-        Error error = new Error.Builder(config, exception, sessionTracker.getCurrentSession(),
+        Error error = new Error.Builder(config, exception, sessionTracker,
             Thread.currentThread(), false)
             .severityReasonType(HandledState.REASON_HANDLED_EXCEPTION)
             .build();
@@ -1007,7 +1002,7 @@ public class Client extends Observable implements Observer {
                                @NonNull StackTraceElement[] stacktrace,
                                @Nullable Callback callback) {
         Error error = new Error.Builder(config, name, message,
-            stacktrace, sessionTracker.getCurrentSession(), Thread.currentThread())
+            stacktrace, sessionTracker, Thread.currentThread())
             .severityReasonType(HandledState.REASON_HANDLED_EXCEPTION)
             .build();
         notify(error, DeliveryStyle.SAME_THREAD, callback);
@@ -1023,7 +1018,7 @@ public class Client extends Observable implements Observer {
     @Deprecated
     public void notifyBlocking(@NonNull Throwable exception,
                                @NonNull MetaData metaData) {
-        Error error = new Error.Builder(config, exception, sessionTracker.getCurrentSession(),
+        Error error = new Error.Builder(config, exception, sessionTracker,
             Thread.currentThread(), false)
             .severityReasonType(HandledState.REASON_HANDLED_EXCEPTION)
             .metaData(metaData)
@@ -1043,7 +1038,7 @@ public class Client extends Observable implements Observer {
     @Deprecated
     public void notifyBlocking(@NonNull Throwable exception, @NonNull Severity severity,
                                @NonNull MetaData metaData) {
-        Error error = new Error.Builder(config, exception, sessionTracker.getCurrentSession(),
+        Error error = new Error.Builder(config, exception, sessionTracker,
             Thread.currentThread(), false)
             .metaData(metaData)
             .severity(severity)
@@ -1070,7 +1065,7 @@ public class Client extends Observable implements Observer {
                                @NonNull Severity severity,
                                @NonNull MetaData metaData) {
         Error error = new Error.Builder(config, name, message,
-            stacktrace, sessionTracker.getCurrentSession(), Thread.currentThread())
+            stacktrace, sessionTracker, Thread.currentThread())
             .severity(severity)
             .metaData(metaData)
             .build();
@@ -1098,7 +1093,7 @@ public class Client extends Observable implements Observer {
                                @NonNull Severity severity,
                                @NonNull MetaData metaData) {
         Error error = new Error.Builder(config, name, message,
-            stacktrace, sessionTracker.getCurrentSession(), Thread.currentThread())
+            stacktrace, sessionTracker, Thread.currentThread())
             .severity(severity)
             .metaData(metaData)
             .build();
@@ -1115,7 +1110,7 @@ public class Client extends Observable implements Observer {
      */
     public void notifyBlocking(@NonNull Throwable exception, @NonNull Severity severity) {
         Error error = new Error.Builder(config, exception,
-            sessionTracker.getCurrentSession(), Thread.currentThread(), false)
+            sessionTracker, Thread.currentThread(), false)
             .severity(severity)
             .build();
         notify(error, BLOCKING);
@@ -1144,7 +1139,7 @@ public class Client extends Observable implements Observer {
 
         @SuppressWarnings("WrongConstant")
         Error error = new Error.Builder(config, exception,
-            sessionTracker.getCurrentSession(), Thread.currentThread(), false)
+            sessionTracker, Thread.currentThread(), false)
             .severity(Severity.fromString(severity))
             .severityReasonType(severityReason)
             .attributeValue(logLevel)
@@ -1302,7 +1297,7 @@ public class Client extends Observable implements Observer {
                         @HandledState.SeverityReason String severityReason,
                         @Nullable String attributeValue, Thread thread) {
         Error error = new Error.Builder(config, exception,
-            sessionTracker.getCurrentSession(), thread, true)
+            sessionTracker, thread, true)
             .severity(severity)
             .metaData(metaData)
             .severityReasonType(severityReason)